Added support for Snapshots (#86)
* #79 Added support for Snapshots

Removed the examples module and added this tests to the integration-tests module.
Added extended features of the example's model to the shared model.

In the tests the injects were removed, because in Integration-Tests, this feature is not available. Instead a Webservice API was built and used

(cherry picked from commit 1d2c9604468a20e61e73c1d741f4692b941bedf3)